Struct asset_transfer_rs::TransferContract

pub struct TransferContract {}

Structure for the AssetContract, on which implemenation transaction functions will be added

Methods

impl TransferContract

The contract implementation Should be marked with the macro #[contrant_impl]

pub fn new() -> TransferContract

pub fn issue_asset(
&self,
asset_id: String,
public_desc: String,
properties: String
) -> Result<(), ContractError>

pub fn invoke_issue_asset(&self, args: Vec<String>) -> Result<(), ContractError>

pub fn agree_to_sell(&self, asset_id: String) -> Result<(), ContractError>

pub fn invoke_agree_to_sell(
&self,
args: Vec<String>
) -> Result<(), ContractError>

pub fn agree_to_buy(&self, asset_id: String) -> Result<(), ContractError>

pub fn invoke_agree_to_buy(
&self,
args: Vec<String>
) -> Result<(), ContractError>

pub fn verify_asset_properties(
&self,
asset_id: String
) -> Result<(), ContractError>

pub fn invoke_verify_asset_properties(
&self,
args: Vec<String>
) -> Result<(), ContractError>

pub fn transfer_asset(&self, asset_id: String) -> Result<(), ContractError>

pub fn invoke_transfer_asset(
&self,
args: Vec<String>
) -> Result<(), ContractError>

pub fn asset_exists(&self, asset_id: String) -> Result<bool, ContractError>

Does the Asset with the supplied key exist

Returns true or false.

pub fn invoke_asset_exists(
&self,
args: Vec<String>
) -> Result<bool, ContractError>

pub fn create_asset(
&self,
my_assset_id: String,
value: String
) -> Result<(), ContractError>

Creates an asset

As an example the value is passed as private data

pub fn invoke_create_asset(
&self,
args: Vec<String>
) -> Result<(), ContractError>

pub fn read_asset_value(
&self,
my_assset_id: String
) -> Result<String, ContractError>

reads an asset and returns the value

pub fn invoke_read_asset_value(
&self,
args: Vec<String>
) -> Result<String, ContractError>